Refactor queries/mutations to separate modules
